feature/routing initial setup
add routing to 4200/records, change app.component.html accordingly

From the tour of heroes course:

There are new requirements for the Tour of Heroes / Tour of Records app:

Add a Dashboard view.
Add the ability to navigate between the Records and Dashboard views.
When users click a recordname in either view, navigate to a detail view of the selected record.
When users click a deep link in an email, open the detail view for a particular record.